moving launch back to normal load for testing
